refactor: Moves core storage components into their own modules
