+++ cmd/aries-agent-rest/go.mod
@@ -4,12 +4,16 @@
 
 module github.com/hyperledger/aries-framework-go/cmd/aries-agent-rest
 
-replace github.com/hyperledger/aries-framework-go => ../..
+replace (
+	github.com/hyperledger/aries-framework-go => ../..
+	github.com/hyperledger/aries-framework-go/component/storage/leveldb => ../../component/storage/leveldb
+)
 
 require (
 	github.com/cenkalti/backoff/v4 v4.1.0
 	github.com/gorilla/mux v1.7.3
 	github.com/hyperledger/aries-framework-go v0.1.5-0.20201017112511-5734c20820a9
+	github.com/hyperledger/aries-framework-go/component/storage/leveldb v0.0.0-00010101000000-000000000000
 	github.com/rs/cors v1.7.0
 	github.com/spf13/cobra v1.0.0
 	github.com/spf13/pflag v1.0.5 // indirect
