+++ component/storage/jsindexeddb/go.mod
@@ -1,0 +1,8 @@
+module github.com/hyperledger/aries-framework-go/component/storage/jsindexeddb
+
+go 1.15
+
+require (
+	github.com/hyperledger/aries-framework-go v0.1.5-0.20201017112511-5734c20820a9
+	github.com/stretchr/testify v1.6.1
+)
