+++ pkg/framework/aries/default_srv.go
@@ -1,5 +1,3 @@
-// +build !js,!wasm
-
 /*
 Copyright SecureKey Technologies Inc. All Rights Reserved.
 
@@ -10,20 +8,9 @@ package aries
 
 import (
 	"github.com/hyperledger/aries-framework-go/pkg/storage"
-	"github.com/hyperledger/aries-framework-go/pkg/storage/leveldb"
 	"github.com/hyperledger/aries-framework-go/pkg/storage/mem"
 )
 
-//nolint:gochecknoglobals
-var (
-	// DBPath Level DB Path.
-	dbPath = "/tmp/peerstore/"
-)
-
-func storeProvider() (storage.Provider, error) {
-	return leveldb.NewProvider(dbPath), nil
-}
-
-func protocolStateStoreProvider() (storage.Provider, error) {
-	return mem.NewProvider(), nil
+func storeProvider() storage.Provider {
+	return mem.NewProvider()
 }
