+++ pkg/framework/aries/framework_test.go
@@ -1,5 +1,3 @@
-// +build !js,!wasm
-
 /*
 Copyright SecureKey Technologies Inc. All Rights Reserved.
 
@@ -49,7 +47,7 @@ import (
 	locallock "github.com/hyperledger/aries-framework-go/pkg/secretlock/local"
 	"github.com/hyperledger/aries-framework-go/pkg/secretlock/local/masterlock/hkdf"
 	"github.com/hyperledger/aries-framework-go/pkg/secretlock/noop"
-	"github.com/hyperledger/aries-framework-go/pkg/storage/leveldb"
+	"github.com/hyperledger/aries-framework-go/pkg/storage/mem"
 	"github.com/hyperledger/aries-framework-go/pkg/vdr/peer"
 )
 
@@ -75,10 +73,6 @@ const doc = `{
 
 func TestFramework(t *testing.T) {
 	t.Run("test framework new - returns error", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		// framework new - error
 		_, err := New(func(opts *Aries) error {
 			return errors.New("error creating the framework option")
@@ -87,10 +81,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test framework new - with default outbound dispatcher", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		// prepare http server
 		server := startMockServer(t, mockHTTPHandler{})
 		port := getServerPort(server)
@@ -135,10 +125,6 @@ func TestFramework(t *testing.T) {
 
 	// framework new - success
 	t.Run("test vdr - with user provided", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		vdr := &mockvdr.MockVDR{}
 		aries, err := New(WithVDR(vdr), WithInboundTransport(&mockInboundTransport{}))
 		require.NoError(t, err)
@@ -159,10 +145,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test vdr - close error", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		vdr := &mockvdr.MockVDR{CloseErr: fmt.Errorf("close vdr error")}
 		aries, err := New(WithVDR(vdr), WithInboundTransport(&mockInboundTransport{}))
 		require.NoError(t, err)
@@ -174,8 +156,9 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test vdr - with default vdr", func(t *testing.T) {
+		t.Skip()
 		// store peer DID in the store
-		dbprov := leveldb.NewProvider(dbPath)
+		dbprov := mem.NewProvider()
 		peerDID := "did:peer:21tDAKCERh95uGgKbJNHYp"
 		store, err := peer.New(dbprov)
 		require.NoError(t, err)
@@ -257,10 +240,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test error from protocol service", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		newMockSvc := func(prv api.Provider) (dispatcher.ProtocolService, error) {
 			return nil, errors.New("error creating the protocol")
 		}
@@ -270,28 +249,16 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test Inbound transport - with options", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		aries, err := New(WithInboundTransport(&mockInboundTransport{}))
 		require.NoError(t, err)
 		require.NotEmpty(t, aries)
 	})
 
 	t.Run("test Inbound transport - start/stop error", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		// start error
 		_, err := New(WithInboundTransport(&mockInboundTransport{startError: errors.New("start error")}))
 		require.Error(t, err)
 		require.Contains(t, err.Error(), "inbound transport start failed")
-
-		path, cleanup = generateTempDir(t)
-		defer cleanup()
-		dbPath = path
 
 		// stop error
 		aries, err := New(WithInboundTransport(&mockInboundTransport{stopError: errors.New("stop error")}))
@@ -304,10 +271,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test KMS svc - with user provided instance", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		// with custom KMS
 		aries, err := New(WithInboundTransport(&mockInboundTransport{}),
 			WithKMS(func(ctx kms.Provider) (kms.KeyManager, error) {
@@ -551,9 +514,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test protocol state store - with user provided protocol state store", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
 		s := storage.NewMockStoreProvider()
 
 		aries, err := New(WithInboundTransport(&mockInboundTransport{}), WithProtocolStateStoreProvider(s))
@@ -563,10 +523,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test new with outbound transport service", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		aries, err := New(WithOutboundTransports(&didcomm.MockOutboundTransport{ExpectedResponse: "data"},
 			&didcomm.MockOutboundTransport{ExpectedResponse: "data1"}))
 		require.NoError(t, err)
@@ -591,10 +547,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test new with transport return route", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		transportReturnRoute := decorator.TransportReturnRouteAll
 		aries, err := New(WithTransportReturnRoute(transportReturnRoute))
 		require.NoError(t, err)
@@ -619,10 +571,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test message service provider option", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		// custom message service provider
 		handler := msghandler.NewMockMsgServiceProvider()
 		aries, err := New(WithMessageServiceProvider(handler))
@@ -636,10 +584,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test default message service provider option", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		// default message service provider
 		aries, err := New()
 		require.NoError(t, err)
@@ -648,10 +592,6 @@ func TestFramework(t *testing.T) {
 	})
 
 	t.Run("test verifiable store option", func(t *testing.T) {
-		path, cleanup := generateTempDir(t)
-		defer cleanup()
-		dbPath = path
-
 		mockStore := &verifiableStoreMocks.MockStore{}
 		// default message service provider
 		aries, err := New(WithVerifiableStore(mockStore))
@@ -685,20 +625,6 @@ func Test_Packager(t *testing.T) {
 		require.Nil(t, f)
 		require.Contains(t, err.Error(), "error from fallback packer")
 	})
-}
-
-func generateTempDir(t testing.TB) (string, func()) {
-	path, err := ioutil.TempDir("", "db")
-	if err != nil {
-		t.Fatalf("Failed to create leveldb directory: %s", err)
-	}
-
-	return path, func() {
-		err := os.RemoveAll(path)
-		if err != nil {
-			t.Fatalf("Failed to clear leveldb directory: %s", err)
-		}
-	}
 }
 
 func startMockServer(t *testing.T, handler http.Handler) net.Listener {
